Update BO architecture in product settings page
