+++ src/PrestaShopBundle/Form/Admin/Configure/ShopParameters/ProductPreferences/StockFormDataProvider.php
@@ -60,6 +60,36 @@ class StockFormDataProvider implements FormDataProviderInterface
      */
     public function setData(array $data)
     {
+        if ($errors = $this->validate($data)) {
+            return $errors;
+        }
+
         return $this->configuration->updateConfiguration($data);
+    }
+
+     /**
+     * Perform validation on form data before saving it.
+     *
+     * @param array $data
+     *
+     * @return array Returns array of errors
+     */
+    protected function validate(array $data)
+    {
+        $errors = [];
+        $displayLastQuantities = $data['display_last_quantities'];
+        if (!is_numeric($displayLastQuantities) || 0 > $displayLastQuantities) {
+            $errors[] = [
+                'key' => 'The %s field is invalid.',
+                'domain' => 'Admin.Notifications.Error',
+                'parameters' => [$this->translator->trans(
+                    'Display remaining quantities when the quantity is lower than',
+                    [],
+                    'Admin.Shopparameters.Feature'
+                )],
+            ];
+        }
+
+        return $errors;
     }
 }
