+++ src/PrestaShopBundle/Resources/views/Admin/Configure/ShopParameters/Blocks/product_preferences_stock.html.twig
@@ -36,6 +36,22 @@
       <div class="card-block row">
         <div class="card-text">
           <div class="form-group row">
+            <label class="form-control-label">
+              {{ 'Enable stock management'|trans }}
+            </label>
+            <div class="col-sm">
+              {{ form_errors(stockForm.stock_management) }}
+              {{ form_widget(stockForm.stock_management) }}
+            </div>
+          </div>
+          <div class="form-group row">
+            {{ ps.label_with_help(('Default pack stock management'|trans), ('When selling packs of products, how do you want your stock to be calculated?'|trans({}, 'Admin.Shopparameters.Help'))) }}
+            <div class="col-sm">
+              {{ form_errors(stockForm.pack_stock_management) }}
+              {{ form_widget(stockForm.pack_stock_management) }}
+            </div>
+          </div>
+          <div class="form-group row">
             {{ ps.label_with_help(('Allow ordering of out-of-stock products'|trans), ('By default, the "%add_to_cart_label%" button is hidden when a product is unavailable. You can choose to have it displayed in all cases.'|trans({'%add_to_cart_label%': 'Add to cart'|trans({}, 'Shop.Theme.Actions')}, 'Admin.Shopparameters.Help'))) }}
             <div class="col-sm">
               {{ form_errors(stockForm.allow_ordering_oos) }}
@@ -43,12 +59,17 @@
             </div>
           </div>
           <div class="form-group row">
-            <label class="form-control-label">
-              {{ 'Enable stock management'|trans }}
-            </label>
+            {{ ps.label_with_help(('Display unavailable attributes on the product page'|trans), ('If an attribute is not available in every product combination, it will not be displayed.'|trans({}, 'Admin.Shopparameters.Help'))) }}
             <div class="col-sm">
-              {{ form_errors(stockForm.stock_management) }}
-              {{ form_widget(stockForm.stock_management) }}
+              {{ form_errors(stockForm.display_unavailable_attributes) }}
+              {{ form_widget(stockForm.display_unavailable_attributes) }}
+            </div>
+          </div>
+          <div class="form-group row">
+            {{ ps.label_with_help(('Display remaining quantities when the quantity is lower than'|trans), ('Set to "0" to disable this feature.'|trans({}, 'Admin.Shopparameters.Help'))) }}
+            <div class="col-sm">
+              {{ form_errors(stockForm.display_last_quantities) }}
+              {{ form_widget(stockForm.display_last_quantities) }}
             </div>
           </div>
           <div class="form-group row">
@@ -92,13 +113,6 @@
               {{ form_errors(stockForm.oos_delivery_time) }}
               {{ form_widget(stockForm.oos_delivery_time) }}
               <small class="form-text">{{ 'Leave empty to disable'|trans }}</small>
-            </div>
-          </div>
-          <div class="form-group row">
-            {{ ps.label_with_help(('Default pack stock management'|trans), ('When selling packs of products, how do you want your stock to be calculated?'|trans({}, 'Admin.Shopparameters.Help'))) }}
-            <div class="col-sm">
-              {{ form_errors(stockForm.pack_stock_management) }}
-              {{ form_widget(stockForm.pack_stock_management) }}
             </div>
           </div>
           {{ form_row(stockForm.oos_show_label_listing_pages) }}
